Exported and renamed a few bundles in the Homematic binding. See openhab/static-code-analysis#100
