+++ addons/binding/org.openhab.binding.homematic/META-INF/MANIFEST.MF
@@ -40,8 +40,12 @@ Import-Package: com.google.common.collect,
  org.eclipse.smarthome.core.thing.type,
  org.eclipse.smarthome.core.types,
  org.jupnp.model.meta,
+ org.openhab.binding.homematic,
+ org.openhab.binding.homematic.handler,
  org.osgi.framework,
  org.osgi.service.component,
  org.osgi.service.component.annotations;resolution:=optional,
  org.slf4j
+Export-Package: org.openhab.binding.homematic,
+ org.openhab.binding.homematic.handler
 Service-Component: OSGI-INF/*.xml
