+++ addons/binding/org.openhab.binding.homematic/src/main/java/org/openhab/binding/homematic/internal/discovery/HomegearDiscoveryParticipant.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.homematic.discovery;
+package org.openhab.binding.homematic.internal.discovery;
 
 import static org.openhab.binding.homematic.HomematicBindingConstants.THING_TYPE_BRIDGE;
