+++ addons/binding/org.openhab.binding.homematic/src/main/java/org/openhab/binding/homematic/internal/discovery/HomematicDeviceDiscoveryService.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.homematic.discovery;
+package org.openhab.binding.homematic.internal.discovery;
 
 import static org.openhab.binding.homematic.HomematicBindingConstants.BINDING_ID;
 
@@ -21,7 +21,7 @@ import org.eclipse.smarthome.core.thing.ThingUID;
 import org.openhab.binding.homematic.handler.HomematicBridgeHandler;
 import org.openhab.binding.homematic.internal.communicator.HomematicGateway;
 import org.openhab.binding.homematic.internal.model.HmDevice;
-import org.openhab.binding.homematic.type.UidUtils;
+import org.openhab.binding.homematic.internal.type.UidUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
