+++ addons/binding/org.openhab.binding.homematic/src/main/java/org/openhab/binding/homematic/internal/type/generator/CcuMetadataExtractor.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.homematic.type.generator;
+package org.openhab.binding.homematic.internal.type.generator;
 
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
